+++ doc-examples/src/main/java/react4j/examples/arez/step2/Footer.java
@@ -3,28 +3,19 @@ package react4j.examples.arez.step2;
 import arez.annotations.Computed;
 import javax.annotation.Nullable;
 import react4j.ReactNode;
-import react4j.annotations.Callback;
 import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
-import react4j.dom.events.MouseEventHandler;
 import react4j.dom.proptypes.html.AnchorProps;
 import react4j.dom.proptypes.html.BtnProps;
 import react4j.dom.proptypes.html.HtmlProps;
 import react4j.examples.arez.step2.model.AppData;
 import react4j.examples.arez.step2.model.FilterMode;
 import static react4j.dom.DOM.*;
-import static react4j.examples.arez.step2.Footer_.*;
 
 @ReactComponent
 abstract class Footer
   extends ReactArezComponent
 {
-  @Callback( MouseEventHandler.class )
-  void handleClearCompleted()
-  {
-    AppData.service.clearCompleted();
-  }
-
   @Nullable
   @Override
   protected ReactNode render()
@@ -48,7 +39,7 @@ abstract class Footer
                   )
               ),
               hasCompletedItems() ?
-              button( new BtnProps().className( "clear-completed" ).onClick( _handleClearCompleted( this ) ),
+              button( new BtnProps().className( "clear-completed" ).onClick( e -> AppData.service.clearCompleted() ),
                       "Clear Completed" ) :
               null
       );