+++ processor/src/main/java/react4j/processor/Constants.java
@@ -8,9 +8,7 @@ final class Constants
   static final String DAGGER_SUBCOMPONENT_CLASSNAME = "dagger.Subcomponent";
   static final String DAGGER_MODULE_CLASSNAME = "dagger.Module";
   static final String DAGGER_PROVIDES_CLASSNAME = "dagger.Provides";
-  static final String JS_FUNCTION_CLASSNAME = "jsinterop.annotations.JsFunction";
   static final String COMPONENT_CLASSNAME = "react4j.Component";
-  static final String CALLBACK_ANNOTATION_CLASSNAME = "react4j.annotations.Callback";
   static final String PROP_ANNOTATION_CLASSNAME = "react4j.annotations.Prop";
   static final String PROP_DEFAULT_ANNOTATION_CLASSNAME = "react4j.annotations.PropDefault";
   static final String PROP_VALIDATE_ANNOTATION_CLASSNAME = "react4j.annotations.PropValidate";