+++ processor/src/test/resources/input/com/example/nested/NestedCompleteComponent.java
@@ -2,7 +2,6 @@ package com.example.nested;
 
 import react4j.Component;
 import react4j.ReactNode;
-import react4j.annotations.Callback;
 import react4j.annotations.Feature;
 import react4j.annotations.Prop;
 import react4j.annotations.ReactComponent;
@@ -15,11 +14,6 @@ public class NestedCompleteComponent
   {
     @Prop
     abstract String getMyProp();
-
-    @Callback
-    void myEventHandler()
-    {
-    }
 
     @Override
     protected ReactNode render()